refactor(radio): move meshtastic client closing logic to MeshtasticBackend
Moved `_close_meshtastic_client` function from main.py to MeshtasticBackend class to improve code decoupling and adhere to the backend abstraction design. Updated tests and documentation to reflect the change.
